Create Concept-file

This function creates a concepts-file in the DDF-format

Usage

create_concepts(
  dataset,
  variable_id,
  variable_name,
  entity_id,
  description = NULL,
  source = NULL,
  name_short = NULL,
  name_catalog = NULL,
  tag_id = NULL,
  output_folder = getwd(),
  return_output = FALSE
)

Arguments

dataset

A data frame. The data set to be converted. Data has to be in wide-format.

variable_id

A character variable. The ID connected to each variable. Should only include lower case letters and underscores.

variable_name

A character variable. The Name of each variable.

entity_id

A character variable. The ID connected to each entity. Should only include lower case letters and underscores.

description

(Optional) A character variable. The definition of each variable. Should not include commas or semi-colons.

source

(Optional) A character variable. The source of each variable. Preferably the URL of the source.

name_short

(Optional) A character variable. A shortened version of the variable name.

name_catalog

(Optional) A character variable. The variable name to be shown in the catalog / tree menu.

tag_id

(Optional) A character variable. The ID specifying which folder the variable belongs in. Should only include lower case letters and underscores.

output_folder

(Optional) Where the file should be saved. By default set to the working directory.

return_output

Returns the output. Default = FALSE

Value

Saves a CSV-file in the given output-folder named "ddf–concepts". This file includes one row for each named variable which always includes geo, year, name, domain, entity_id, all data-variables and the given, relevant variables.

Examples

if (FALSE) { # \dontrun{
create_concepts(dataset = dataset, variable_id = "var_id", variable_name = "Variable",
entity_id = "reg_id", source = "source_url")
} # }